Fix: wrong expiration time displayed on dashboard
